Fixed some help popup problems
